django_army/requirements.txt
Django==1.8.14
virtualenv==15.1.0
django-model-utils==1.4.0